backend/services/odds-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
aiohttp==3.9.0
